refactor: move type
Move type that is only used in one file into that file.
