+++ src/services/resolve-remote-packument-version.ts
@@ -3,16 +3,35 @@ import { Registry } from "../domain/registry";
 import { AsyncResult, Err } from "ts-results-es";
 import {
   ResolvableVersion,
-  ResolvedPackumentVersion,
   ResolvePackumentVersionError,
 } from "../packument-version-resolving";
 import { PackumentNotFoundError } from "../common-errors";
 import {
   tryResolvePackumentVersion,
   UnityPackument,
+  UnityPackumentVersion,
 } from "../domain/packument";
 import { FetchPackument } from "../io/packument-io";
 import { resultifyAsyncOp } from "../utils/result-utils";
+import { RegistryUrl } from "../domain/registry-url";
+
+/**
+ * A successfully resolved packument-version.
+ */
+export interface ResolvedPackumentVersion {
+  /**
+   * The packument from which the version was resolved.
+   */
+  readonly packument: UnityPackument;
+  /**
+   * The resolved packument-version.
+   */
+  readonly packumentVersion: UnityPackumentVersion;
+  /**
+   * The source from which the packument was resolved.
+   */
+  readonly source: RegistryUrl;
+}
 
 /**
  * Function for resolving remove packument versions.
